Add separate flags for composite types and templates for DM create.
Also, add a warning when using '--config' with a template instead of the new '--template' flag.  Warn users that this will be deprecated 2017/11/08.

----Release Notes----

Add separate flags for composite types and templates for DM create.

Also, add a warning when using '--config' with a template instead of the new '--template' flag.  Warn users that this will be deprecated 2017/11/08.

----Internal Notes----

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164653352