Configure Shared VPC when creating a project
This adds the possibility to configure the Shared VPC
when creating a project. A project can be the
host of a Shared VPC or can be linked to another
project to use its Shared VPC (it's then a "service
project").

This adds a specific config file as example for this
feature.

This also fixes a problem of resource name conflict