Deployment Manager Shell Scripts to deploy a 3 tier web app using jinja templates. (#140)
This is sample automation using shell script and jinja templates to deploy a web app comprised of two microservices. An instance template is used ot generate and instance group for the back-end microservice, which is then fronted by an internal TCP load balancer, with autoscaling and health-check. Similarly, the front-end microservice is created in the same way but fronted by an HTTP load balancer. Firewall rules are created to allow inter-microservice traffic as well as enable external acces and health-checks etc.