Shared VPC projects can be created in a folder
Fix #351

We didn't allow the user to specify both an org ID and a folder ID,
but both are needed to create Shared VPC projects in a folder.

If both an org ID and a folder ID are provided, the project will
be created in the folder and not in at the org root.