Amend existing GCS bucket templates to allow domain-named buckets
This template makes use of references, which require the names of
the resources being referenced to not include a '.' character in
the name. Because the resource name and the bucket name have been
one in the same for this template, a domain-named bucket causes
the deployment to fail because the reference cannot be resolved.

To fix the issue, this separates the name of the bucket (with the
domain) from the name of the resource, allowing domain-named
buckets while still preserving the references being used throughout
the template.