Update BigQuery template to include GET query parameter
BigQuery recently added a new query parameter to their tables.get
method that required us to add a new property that maps to it. To
avoid conflict with the existing table `view` property, users can
now specify the `tableMetadataView` property on bigquery table
resources which will control what is returned on GET requests.